+++ src/ripple_detection/detectors/_long.py
@@ -17,6 +17,7 @@ from ripple_detection.core import (
 )
 from ripple_detection.detectors._blocks import (
     _drop_short_blocks,
+    _reject_flat_channels,
     _valid_blocks,
 )
 from ripple_detection.detectors._events import (
@@ -252,6 +253,7 @@ def Long_sharp_wave_ripple_detector(
     )
     n_time = len(time)
     is_valid, blocks = _valid_blocks(time, lfp)
+    _reject_flat_channels(lfp, is_valid, "raw_lfps")
     slowest_kernel = len(_gaussian_lowpass_fir(sharp_wave_band[0], sampling_frequency))
     blocks = _drop_short_blocks(
         blocks,